Add config parameter to preserve script closures
This commit will keep mars-game working when the event-handler replication branch is merged to development (but keeps us from taking advantage of event handler replication).  We will eventually need to eliminate our use of closure variables in methods and event handlers so that we can remove this config parameter and take advantage of event handler replication.