Push new triggers to a local reference and reassign that reference.
Properties may now contain node references in arrays and other objects.
When those properties are referenced, the nodes are translated from the
kernel-format ID to the component's node reference. Since part of the
array changes, the containing array is also a different object so that
it may contain the translated value(s).

To modify an object containing property references, the translated
property must be retained locally, modified, then reassigned to the
property. If a local copy is not retained and reassigned, the intended
change will happen to the temporary version of the array containing the
translated value, but the assignment will read and assign the original
array.

`canFire$` changes to match `triggers` even though it doesn't contain
property references and doesn't strictly need to change.